+++ src/test/java/software/amazon/encryption/s3/S3EncryptionClientTest.java
@@ -171,36 +171,6 @@ public class S3EncryptionClientTest {
     }
 
     @Test
-    public void defaultModeWithLargeObjectFails() throws IOException {
-        final String objectKey = "large-object";
-
-        // V3 Client
-        S3Client v3Client = S3EncryptionClient.builder()
-                .aesKey(AES_KEY)
-                .build();
-
-        // Tight bound on the default limit of 64MiB
-        final long fileSizeExceedingDefaultLimit = 1024 * 1024 * 64 + 1;
-        final InputStream largeObjectStream = new BoundedZerosInputStream(fileSizeExceedingDefaultLimit);
-        v3Client.putObject(PutObjectRequest.builder()
-                .bucket(BUCKET)
-                .key(objectKey)
-                .build(), RequestBody.fromInputStream(largeObjectStream, fileSizeExceedingDefaultLimit));
-
-        largeObjectStream.close();
-
-        assertThrows(S3EncryptionClientException.class, () -> v3Client.getObjectAsBytes(builder -> builder
-                .bucket(BUCKET)
-                .key(objectKey)));
-
-
-        // Cleanup
-        deleteObject(BUCKET, objectKey, v3Client);
-        v3Client.close();
-    }
-
-
-    @Test
     public void s3EncryptionClientWithKeyringFromKmsKeyIdSucceeds() {
         final String objectKey = "keyring-from-kms-key-id";
